Create function that assigns operands and operator
Create function for each number button event listener, that logs the
first operator variable and second operator variable, and keeps the
display of each independent of each other.

The division event listener function is also defined, assigning the
operator variable to '/', and confirming the correct process in
assigning and displaying each of the three variables for the operate
function.