Update decimal and zero button event listeners
Make zero button event listener assign variable zero/display zero using
equals operator when conditions of first and or second operand are
undefined (depending on if zero button event listener is called
before or after operator) OR equal to zero. This has the effect of not
allowing multiple zeros to populate the display or be assigned to a
variable, unless following a 1-9 digit.

Similarly, for the decimal button, I adjusted functions so that decimals
can only be added to variables/display if there is not one already
present.